# swagger_server/models/element.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Element(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: str=None):  # noqa: E501
        """Element - a model defined in Swagger

        :param id: The id of this Element.  # noqa: E501
        :type id: str
        """
        self.swagger_types = {
            'id': str
        }

        self.attribute_map = {
            'id': 'id'
        }

        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> 'Element':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Element of this Element.  # noqa: E501
        :rtype: Element
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Element.


        :return: The id of this Element.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Element.


        :param id: The id of this Element.
        :type id: str
        """
        allowed_values = ["amcl", "kinect", "mrpt", "lidar", "aruco", "camera"]  # noqa: E501
        if id not in allowed_values:
            raise ValueError(
                "Invalid value for `id` ({0}), must be one of {1}"
                .format(id, allowed_values)
            )

        self._id = id