# swagger_server/models/parameters.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Parameters(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, error: str=None):  # noqa: E501
        """Parameters - a model defined in Swagger

        :param error: The error of this Parameters.  # noqa: E501
        :type error: str
        """
        self.swagger_types = {
            'error': str
        }

        self.attribute_map = {
            'error': 'error'
        }

        self._error = error

    @classmethod
    def from_dict(cls, dikt) -> 'Parameters':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Parameters of this Parameters.  # noqa: E501
        :rtype: Parameters
        """
        return util.deserialize_model(dikt, cls)

    @property
    def error(self) -> str:
        """Gets the error of this Parameters.

        human readable text describing the error, if possible. (TODO: add error codes here as we discover them)  # noqa: E501

        :return: The error of this Parameters.
        :rtype: str
        """
        return self._error

    @error.setter
    def error(self, error: str):
        """Sets the error of this Parameters.

        human readable text describing the error, if possible. (TODO: add error codes here as we discover them)  # noqa: E501

        :param error: The error of this Parameters.
        :type error: str
        """
        if error is None:
            raise ValueError("Invalid value for `error`, must not be `None`")  # noqa: E501

        self._error = error